+++ flowauth/frontend/cypress/integration/login_spec.js
@@ -12,7 +12,7 @@ describe("Login screen", function () {
 		cy.get("#username").type("TEST_USER");
 		cy.get("#password").type("DUMMY_PASSWORD");
 		cy.get("button").click();
-		cy.contains("Aruba").should("exist");
+		cy.contains("TEST_SERVER").should("exist");
 		cy.getCookie("session").should("exist");
 		cy.getCookie("X-CSRF").should("exist");
 	});